# docs/swagger.yaml
basePath: /
definitions:
  types.APIResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: string
    type: object
...
  types.CallbackRequest:
    properties:
      code:
        description: 授权码
        example: authorization_code
        minLength: 1
        type: string
      provider:
        description: 平台名称 可选值：youtube x facebook tiktok instagram
        enum:
          - youtube
          - x
          - facebook
          - tiktok
          - instagram
        example: x
        type: string
      redirect_uri:
        description: 重定向URI
        example: hhttps://test-pubproject.wondera.io/static/callback.html
        type: string
      server_name:
        description: 服务器名称
        example: myapp
        maxLength: 50
        minLength: 1
        type: string
      state:
        description: 状态参数，包含用户ID等信息
        example: encoded_state_string
        minLength: 1
        type: string
      user_id:
        description: 服务内部用户ID 必填
        example: user123
        maxLength: 100
        minLength: 1
        type: string
    required:
      - code
      - provider
      - redirect_uri
      - server_name
      - state
      - user_id
    type: object
  types.CallbackResponse:
    properties:
      expires_at:
        description: 时间戳格式
        example: 1704067199
        type: integer
      message:
        example: OAuth callback completed successfully
        type: string
      provider:
        example: x
        type: string
      refer_at:
        description: 时间戳格式
        example: 1704067199
        type: integer
      server_name:
        example: myapp
        type: string
      user_id:
        example: user123
        type: string
    type: object
  types.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      request_id:
        type: string
    type: object
...
  types.GetUserInfoRequest:
    properties:
      provider:
        description: 平台名称
        enum:
          - youtube
          - x
          - facebook
          - tiktok
          - instagram
        example: x
        type: string
      server_name:
        description: 服务名称
        example: myapp
        maxLength: 50
        minLength: 1
        type: string
      user_id:
        description: 用户ID
        example: user123
        maxLength: 100
        minLength: 1
        type: string
    required:
      - provider
      - server_name
      - user_id
    type: object
  types.GetUserInfoResponse:
    properties:
      provider:
        example: x
        type: string
      server_name:
        example: myapp
        type: string
      user_id:
        example: user123
        type: string
      user_info:
        $ref: "#/definitions/types.UserInfo"
    type: object
  types.IsAuthorizedRequest:
    properties:
      provider:
        enum:
          - youtube
          - x
          - facebook
          - tiktok
          - instagram
        example: x
        type: string
      server_name:
        example: myapp
        maxLength: 50
        minLength: 1
        type: string
      user_id:
        example: user123
        maxLength: 100
        minLength: 1
        type: string
    required:
      - provider
      - server_name
      - user_id
    type: object
  types.IsAuthorizedResponse:
    properties:
      is_authorized:
        example: true
        type: boolean
    type: object
...
  types.RefreshTokenRequest:
    properties:
      provider:
        description: 平台名称
        enum:
          - youtube
          - x
          - facebook
          - tiktok
          - instagram
        example: x
        type: string
      server_name:
        description: 服务名称
        example: myapp
        maxLength: 50
        minLength: 1
        type: string
      user_id:
        description: 用户ID
        example: user123
        maxLength: 100
        minLength: 1
        type: string
    required:
      - provider
      - server_name
      - user_id
    type: object
  types.RefreshTokenResponse:
    properties:
      expires_at:
        description: 新token的过期时间戳
        example: 1704067199
        type: integer
      message:
        example: Token refreshed successfully
        type: string
      provider:
        example: x
        type: string
      refreshed_at:
        description: 刷新时间戳
        example: 1704067199
        type: integer
      server_name:
        example: myapp
        type: string
      user_id:
        example: user123
        type: string
    type: object
...
  types.StartAuthRequest:
    properties:
      provider:
        description: 平台名称 可选值：youtube x facebook tiktok instagram
        enum:
          - youtube
          - x
          - facebook
          - tiktok
          - instagram
        example: x
        type: string
      redirect_uri:
        example: https://test-pubproject.wondera.io/static/callback.html
        type: string
      server_name:
        example: myapp
        maxLength: 50
        minLength: 1
        type: string
      user_id:
        description: 用户ID 必填 同一服务名称下user_id唯一
        example: user123
        maxLength: 100
        minLength: 1
        type: string
    required:
      - provider
      - redirect_uri
      - server_name
      - user_id
    type: object
  types.StartAuthResponse:
    properties:
      auth_url:
        example: https://x.com/i/oauth2/authorize?response_type=code&client_id=...
        type: string
      provider:
        example: x
        type: string
      server_name:
        example: myapp
        type: string
      user_id:
        example: user123
        type: string
    type: object
...
  types.UserInfo:
    properties:
      avatar_url:
        description: 头像URL
        example: https://example.com/avatar.jpg
        type: string
      display_name:
        description: 显示名称
        example: John Doe
        type: string
      email:
        description: 邮箱（如果可用）
        example: john@example.com
        type: string
      followers:
        description: 粉丝数
        example: 1000
        type: integer
      following:
        description: 关注数
        example: 500
        type: integer
      id:
        description: 平台用户ID
        example: "1234567890"
        type: string
      profile_url:
        description: 个人资料URL
        example: https://x.com/johndoe
        type: string
      username:
        description: 用户名
        example: johndoe
        type: string
      verified:
        description: 是否认证用户
        example: false
        type: boolean
    type: object
host: localhost:8084
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 多平台社交媒体授权分享API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Social Media Platform API
  version: "1.0"
paths:
...
  /auth/callback:
    post:
      consumes:
        - application/json
      description: 前端收到第三方平台OAuth回调后，调用此接口处理授权码交换和token保存
      parameters:
        - description: 回调请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/types.CallbackRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OAuth callback completed
          schema:
            allOf:
              - $ref: "#/definitions/types.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/types.CallbackResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
      summary: 处理OAuth回调
      tags:
        - 认证
  /auth/is-authorized:
    post:
      consumes:
        - application/json
      description: 查询指定用户是否授权指定平台
      parameters:
        - description: 查询是否授权请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/types.IsAuthorizedRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
              - $ref: "#/definitions/types.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/types.IsAuthorizedResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
      summary: 查询是否授权
      tags:
        - 认证
  /auth/refresh-token:
    post:
      consumes:
        - application/json
      description: 客户端主动刷新指定平台的访问token
      parameters:
        - description: 刷新token请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/types.RefreshTokenRequest"
      produces:
        - application/json
      responses:
        "200":
          description: token刷新成功
          schema:
            allOf:
              - $ref: "#/definitions/types.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/types.RefreshTokenResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "401":
          description: token不存在或刷新失败
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
      summary: 手动刷新token
      tags:
        - 认证
  /auth/start:
    post:
      consumes:
        - application/json
      description: 启动指定平台的OAuth授权流程，返回授权URL
      parameters:
        - description: 授权请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/types.StartAuthRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 授权URL生成成功
          schema:
            allOf:
              - $ref: "#/definitions/types.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/types.StartAuthResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
      summary: 开始OAuth授权流程
      tags:
        - 认证
  /auth/user-info:
    post:
      consumes:
        - application/json
      description: 获取指定平台用户的详细信息
      parameters:
        - description: 获取用户信息请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/types.GetUserInfoRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 获取用户信息成功
          schema:
            allOf:
              - $ref: "#/definitions/types.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/types.GetUserInfoResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "401":
          description: 未授权或token过期
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
      summary: 获取用户信息
      tags:
        - 认证
...
schemes:
  - http
  - https
swagger: "2.0"